pyproject.toml
[project]
name = "gws-cli"
version = "0.1.0"
description = "Read-only CLI for Google Workspace resources"
readme = "README.md"
authors = [
    { name = "TAKAHISA Sakai", email = "sakai.takahisa@classmethod.jp" }
]
requires-python = ">=3.13"
dependencies = [
    "google-api-python-client>=2.194.0",
    "google-auth>=2.49.2",
    "markdownify>=1.2.2",
    "typer>=0.12",
]

[build-system]
requires = ["uv_build>=0.11.6,<0.12.0"]
build-backend = "uv_build"

[project.scripts]
gws-cli = "gws_cli.main:app"

[dependency-groups]
dev = [
    "pytest>=9.0.3",
]
